+++ angular/projects/wcs-angular/src/lib/proxies.ts
@@ -1363,7 +1363,7 @@ import type { WcsTabChangeEvent as IWcsTabsWcsTabChangeEvent } from 'wcs-core';
 
 export declare interface WcsTabs extends Components.WcsTabs {
   /**
-   * 
+   *
 Emitted when the selected tab change.
    */
   tabChange: EventEmitter<CustomEvent<IWcsTabsWcsTabChangeEvent>>;