+++ example/src/app/input-example/input-example.component.ts
@@ -5,10 +5,12 @@ import { Component, OnInit } from '@angular/core';
   template: `
     <h2>Exemple utilisation input</h2>
     <h3>Input bindé avec un ngModel</h3>
-    Input 1
+    Input 1 (L)
+    <wcs-input [(ngModel)]="inputText1" size='l' placeholder="exemple d'input bindé avec ngModel"></wcs-input>
+    Input 2 (M)
     <wcs-input [(ngModel)]="inputText1" placeholder="exemple d'input bindé avec ngModel"></wcs-input>
-    Input 2
-    <wcs-input [(ngModel)]="inputText1" placeholder="exemple d'input bindé avec ngModel"></wcs-input>
+    Input 3 (S)
+    <wcs-input [(ngModel)]="inputText1" size='s' placeholder="exemple d'input bindé avec ngModel"></wcs-input>
 
     <h3>Idem avec textarea</h3>
     Textarea 1